flask-backend/requirements.txt
aniso8601
blinker
boto3
botocore
certifi
cffi
charset-normalizer
click
cryptography
Flask
flask-cors
flask-jwt-extended>=4.0.0
aws-wsgi
Flask-RESTful
Flask-Testing
idna
iniconfig
itsdangerous
Jinja2
jmespath
MarkupSafe
moto
packaging
pluggy
pycparser
PyJWT
pytest
pytest-mock
python-dateutil
python-dotenv
pytz
PyYAML
requests
responses
s3transfer
six
urllib3
Werkzeug
xmltodict
